requirements/dev.txt
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coverage==6.4.1
crispy-bootstrap5==0.6
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.5
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-dotenv==1.4.2
django-redis==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
exceptiongroup==1.0.0rc9
flake8==4.0.1
h11==0.14.0
idna==3.3
iniconfig==1.1.1
install==1.3.5
mccabe==0.6.1
model-bakery==1.9.0
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-django==4.5.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.1
redis==4.3.3
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.5.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.8.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.8.4
wrapt==1.14.1
wsproto==1.2.0
